fix(model): validate default model against available options
